Comments on code
So deleted your delete.rb file, created a method on your car lot for "selling/deleting vehicles"

User selects 3, that is called, and runs.

We build a list of vehicles with an index
User selects from list
If user makes an invalid selection we let them know, and kick them to main menu
If the car is valid, the car is "sold/deleted" removed from the list of vehicles.